pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = "Test project cor code review of repository"
authors = ["shpilievoi29 <shpilevoy29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
email-validator = "^2.2.0"
fastapi = "^0.115.3"
fastapi-pagination = "^0.12.31"
jinja2 = "^3.1.4"
mako = "^1.3.6"
openai = "^1.52.1"
passlib = "^1.7.4"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.12"
ujson = "^5.10.0"
uvicorn = "^0.32.0"
autoflake = "^2.3.1"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
isort = "^5.13.2"
anybadge = "^1.14.0"
coverage = "^7.6.4"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
faker = "^30.8.0"
redis = "^5.1.1"
aioredis = "^2.0.1"
slowapi = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
